Move objects edit css html with flex
I'm having some difficulty getting the list to behave the way I want it to. For example right now I'm struggling to get the link items to display as a row. I've tried setting the flex column to row but it doesn't make a difference. I'll continue to play around with it until I come up with a workable solution.