Move notif bar to the right
After some struggle I've successfully managed to move the notification bar and profile icon to the right side of the nav bar. As it turns out I just needed to define the right section as its own flex container and then use a descendant combinator to margin that section to the right of the header.