Turn .modal into flex container use f.shrink icon
I missed the most simple thing, and that was turning the base div container into a flex container. I did try it earlier but I ran into the problem of it squishing my icon and close button. I fixed this by setting the flex-shrink to 0 on both of these elements.