Define flex container rules for alignment
This one I've figured out fairly quickly overall. Once I set the body to a flex column I was able to add containers for the content on the right side of the page. I then specified the width and height of our box content and wrapped them. I also put a slight margin and gap around the boxes.